% Auto-generated: do not edit by hand
\name{rangeSelector}

\alias{rangeSelector}

\title{RangeSelector component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
rangeSelector(id=NULL, cursor_color=NULL, height=NULL, range=NULL,
selection_color=NULL, width=NULL, xend=NULL, xstart=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{cursor_color}{Character. Color of the cursor line.}

\item{height}{Character. Height of the selection area.}

\item{range}{List of numerics. Range of the selection.}

\item{selection_color}{Character. Color of the selection area.}

\item{width}{Character. Width of the selection area.}

\item{xend}{Numeric. End value of the selection.}

\item{xstart}{Numeric. Start value of the selection.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

